added place holder images to index.ejs
