--- server/views/index.ejs
@@ -1,64 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <meta http-equiv="X-UA-Compatible" content="ie=edge">
-  <title>Movies Now</title>
-  <link rel="stylesheet" type="text/css" href="../../www/css/reset.css" />
-  <link rel="stylesheet" type="text/css" href="../../www/css/base.css" />
-  <link rel="stylesheet" type="text/css" href="../../www/css/layout.css" />
-  <link rel="stylesheet" type="text/css" href="../../www/css/module.css" />
-  <link rel="stylesheet" type="text/css" href="../../www/css/theme.css" />
-</head>
-
-<body>
-  <main>
-
-    <form>
-      <label>
-        <input type="text" name="Location" value="">
-      </label>
-
-      <h1>MovieNow</h1>
-
-      <label>
-        <input type="text" name="Time" Value="">
-      </label>
-
-      <section class="movies">
-        <div class="image">
-          <p> 1 </p>
-        </div>
-        <div class="image">
-          <p> 2 </p>
-        </div>
-        <div class="image">
-          <p> 3 </p>
-        </div>
-        <div class="image">
-          <p> 4 </p>
-        </div>
-        <div class="image">
-          <p> 5 </p>
-        </div>
-        <div class="image">
-          <p> 6 </p>
-        </div>
-      </section>
-
-    </form>
-  </main>
-
-  <footer>
-    <button type="button"> Sign In </button>
-    <button type="button"> Go Now </button>
-    <button type="button"> About Us </button>
-  </footer>
-
-
-
-</body>
-
-</html>
