+++ views/index.ejs
@@ -1,0 +1,68 @@
+<!DOCTYPE html>
+<html lang="en">
+
+
+<%- include('../views/head') %>
+
+
+<body>
+  <main>
+
+    <form>
+      <label>
+        <input type="text" name="Location" value="">
+      </label>
+
+      <h1>MovieNow</h1>
+
+      <label>
+        <input type="text" name="Time" Value="">
+      </label>
+
+      <section class="movies">
+
+        <div class="image">
+          <a href="https://placeholder.com"><img
+              src="https://via.placeholder.com/200x300.JPEG?text=First+Movie+Avaliable"></a>
+        </div>
+
+        <div class="image">
+          <a href="https://placeholder.com">
+            <img src="https://via.placeholder.com/200x300.JPEG?text=Second+Movie+Available"></a>
+        </div>
+
+        <div>
+          <a href="https://placeholder.com">
+            <img src="https://via.placeholder.com/200x300.JPEG?text=Third+Movie+Available"></a>
+        </div>
+
+        <div class=" image">
+          <a href="https://placeholder.com">
+            <img src="https://via.placeholder.com/200x300.JPEG?text=Forth+Movie+Available"></a>
+        </div>
+
+        <div class="image">
+          <a href="https://placeholder.com">
+            <img src="https://via.placeholder.com/200x300.JPEG?text=Fifth+Movie+Available"></a>
+        </div>
+
+        <div class="image">
+          <a href="https://placeholder.com">
+            <img src="https://via.placeholder.com/200x300.JPEG?text=Sixth+Movie+Available"></a>
+        </div>
+
+      </section>
+    </form>
+  </main>
+
+  <footer>
+    <button type="button"> Sign In </button>
+    <button type="button"> Go Now </button>
+    <button type="button"> About Us </button>
+  </footer>
+
+
+
+</body>
+
+</html>
